chore(Refactor): Project Structure Refactor
- Separate files per concern (Components and Pages)
- Add Module alias
