+++ src/components/Display/styles/Display.css
