+++ src/index.tsx
@@ -1,6 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import App from "./App";
+import App from "@pages/App";
+
 import "./index.css";
 
 ReactDOM.render(
